refactor(gnss): move GnssChecks to src/lib/gnss
